Complete Master Definition Backend
